refactor: move canvas into page slice
