+++ src/pages/canvas/widgets/PipelineCanvas/CanvasCodeCard.tsx
@@ -8,8 +8,8 @@ import { useAtom, useAtomValue, useSetAtom } from 'jotai';
 import type React from 'react';
 import { useMemo, useRef } from 'react';
 import { focusedPaneAtom } from '@/entities/AppView/model/atoms';
-import type { CanvasNode } from '../../../entities/CanvasNode/model/types.ts';
-import CodeCard from '../../CodeCard/CodeCard.tsx';
+import type { CanvasNode } from '@/entities/CanvasNode/model/types.ts';
+import CodeCard from '@/widgets/CodeCard/CodeCard.tsx';
 import { cardPositionsAtom, selectedNodeIdsAtom, transformAtom } from './model/atoms.ts';
 
 interface CanvasCodeCardProps {
