+++ src/pages/canvas/widgets/PipelineCanvas/PipelineCanvas.tsx
@@ -4,10 +4,10 @@ import { useEffect, useMemo, useRef } from 'react';
 import { useHotkeys } from 'react-hotkeys-hook';
 // Atoms & Hooks
 import { filesAtom, focusedPaneAtom, fullNodeMapAtom, graphDataAtom } from '@/entities/AppView/model/atoms';
+import CopyAllCodeButton from '@/features/CopyAllCodeButton.tsx';
+import ResetViewButton from '@/features/ResetViewButton.tsx';
 import { symbolMetadataAtom } from '@/features/Search/UnifiedSearch/model/atoms.ts';
 import { extractSymbolMetadata } from '@/shared/symbolMetadataExtractor.ts';
-import CopyAllCodeButton from '../../../features/CopyAllCodeButton.tsx';
-import ResetViewButton from '../../../features/ResetViewButton.tsx';
 import { CanvasCodeCard } from './CanvasCodeCard.tsx';
 import CanvasConnections from './CanvasConnections.tsx';
 import D3ZoomContainer from './D3ZoomContainer.tsx';
