+++ src/pages/canvas/widgets/PipelineCanvas/useCanvasLayout.ts
@@ -1,8 +1,8 @@
 import { useAtomValue, useSetAtom } from 'jotai';
 import { useEffect, useState } from 'react';
 import { fullNodeMapAtom } from '@/entities/AppView/model/atoms';
-import type { CanvasNode } from '../../../entities/CanvasNode/model/types.ts';
-import type { GraphData } from '../../../entities/SourceFileNode/model/types.ts';
+import type { CanvasNode } from '@/entities/CanvasNode/model/types.ts';
+import type { GraphData } from '@/entities/SourceFileNode/model/types.ts';
 import { layoutLinksAtom, layoutNodesAtom } from './model/atoms.ts';
 
 export const useCanvasLayout = (_initialData: GraphData | null, visibleNodeIds: Set<string>) => {
