+++ src/widgets/CodeCard/hooks/useFocusedIdentifiers.ts
@@ -9,8 +9,8 @@ import { useAtom, useAtomValue, useSetAtom } from 'jotai';
 import { useCallback, useMemo } from 'react';
 import { fullNodeMapAtom } from '@/entities/AppView/model/atoms';
 import { activeLocalVariablesAtom } from '@/features/Code/FocusMode/model/atoms';
-import { visibleNodeIdsAtom } from '@/widgets/MainContents/PipelineCanvas/model/atoms';
-import { pruneDetachedNodes } from '@/widgets/MainContents/PipelineCanvas/utils';
+import { visibleNodeIdsAtom } from '@/pages/canvas/widgets/PipelineCanvas/model/atoms';
+import { pruneDetachedNodes } from '@/pages/canvas/widgets/PipelineCanvas/utils';
 import type { CanvasNode } from '../../../entities/CanvasNode/model/types';
 import type { CodeLine } from '../../../entities/CodeLine/model/types';
