+++ src/widgets/CodeViewer/ui/segments/ExpandSegment.tsx
@@ -5,7 +5,7 @@
 import { useAtomValue, useSetAtom } from 'jotai';
 import type React from 'react';
 import { hoveredIdentifierAtom } from '@/entities/AppView/model/atoms';
-import { visibleNodeIdsAtom } from '@/widgets/MainContents/PipelineCanvas/model/atoms';
+import { visibleNodeIdsAtom } from '@/pages/canvas/widgets/PipelineCanvas/model/atoms';
 import type { CanvasNode } from '../../../../entities/CanvasNode/model/types';
 import type { CodeSegment, SegmentStyle } from '../../core/types';
